+++ test/acceptance/test/core_operations_test.go
@@ -1,3 +1,8 @@
+// +build !unittest
+// +build acceptance
+// +build !smoke
+
+// ^^ temporarily disabling this test until we have it in ginkgo format.
 package acceptance
 
 // Runs basic WeGO operations against a kind cluster.
